+++ UnityEditor.TestRunner/TestLaunchers/PlatformSetup/AndroidPlatformSetup.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Net;
+using UnityEditor.Build;
 using UnityEngine;
 
 namespace UnityEditor.TestTools.TestRunner
@@ -57,9 +58,13 @@ namespace UnityEditor.TestTools.TestRunner
 
         public void Setup()
         {
+#if UNITY_2021_1_OR_NEWER            
+            m_oldApplicationIdentifier = PlayerSettings.GetApplicationIdentifier(NamedBuildTarget.Android);
+#else 
             m_oldApplicationIdentifier = PlayerSettings.GetApplicationIdentifier(BuildTargetGroup.Android);
+#endif
             PlayerSettings.SetApplicationIdentifier(BuildTargetGroup.Android, "com.UnityTestRunner.UnityTestRunner");
-
+            
             if (RequiresLegacyConnectionMechanism)
                 PerformLegacySetup();
 
@@ -87,8 +92,11 @@ namespace UnityEditor.TestTools.TestRunner
         {
             if (RequiresLegacyConnectionMechanism)
                 PerformLegacyCleanup();
-
+#if UNITY_2021_1_OR_NEWER 
+            PlayerSettings.SetApplicationIdentifier(NamedBuildTarget.Android, m_oldApplicationIdentifier);
+#else
             PlayerSettings.SetApplicationIdentifier(BuildTargetGroup.Android, m_oldApplicationIdentifier);
+#endif
         }
     }
 }