+++ UnityEditor.TestRunner/TestLaunchers/PlatformSetup/UwpPlatformSetup.cs
@@ -1,4 +1,5 @@
 using System;
+using UnityEditor.Build;
 
 namespace UnityEditor.TestTools.TestRunner
 {
@@ -23,11 +24,20 @@ namespace UnityEditor.TestTools.TestRunner
             // Otherwise we can use the existing configuration specified by the user in Build Settings.
             if (Environment.GetEnvironmentVariable("UNITY_THISISABUILDMACHINE") == "1" || wsaSettingNotInitialized)
             {
+#if UNITY_2023_1_OR_NEWER
+#else 
                 EditorUserBuildSettings.wsaSubtarget = WSASubtarget.PC;
+#endif
+                
                 EditorUserBuildSettings.wsaArchitecture = "x64";
                 EditorUserBuildSettings.SetPlatformSettings(BuildPipeline.GetBuildTargetName(BuildTarget.WSAPlayer), k_SettingsBuildConfiguration, WSABuildType.Debug.ToString());
                 EditorUserBuildSettings.wsaUWPBuildType = WSAUWPBuildType.ExecutableOnly;
+#if UNITY_2021_1_OR_NEWER
+                PlayerSettings.SetIl2CppCompilerConfiguration(NamedBuildTarget.WindowsStoreApps, Il2CppCompilerConfiguration.Debug);
+#else
                 PlayerSettings.SetIl2CppCompilerConfiguration(BuildTargetGroup.WSA, Il2CppCompilerConfiguration.Debug);
+#endif
+
             }
         }