+++ UnityEditor.TestRunner/TestRun/Tasks/BuildTestTreeTask.cs
@@ -18,7 +18,7 @@ namespace UnityEditor.TestTools.TestRunner.TestRun.Tasks
         }
 
         internal IEditorLoadedTestAssemblyProvider m_testAssemblyProvider = new EditorLoadedTestAssemblyProvider(new EditorCompilationInterfaceProxy(), new EditorAssembliesProxy());
-        internal Func<string[], IAsyncTestAssemblyBuilder> m_testAssemblyBuilderFactory = orderedTestNames => new UnityTestAssemblyBuilder(orderedTestNames);
+        internal Func<string[], int, IAsyncTestAssemblyBuilder> m_testAssemblyBuilderFactory = (orderedTestNames, seed) => new UnityTestAssemblyBuilder(orderedTestNames, seed);
         internal ICallbacksDelegator m_CallbacksDelegator = CallbacksDelegator.instance;
 
         public override IEnumerator Execute(TestJobData testJobData)
@@ -41,7 +41,7 @@ namespace UnityEditor.TestTools.TestRunner.TestRun.Tasks
 
             var assemblies = assembliesEnumerator.Current.Where(pair => m_TestPlatform.IsFlagIncluded(pair.Key)).SelectMany(pair => pair.Value).Select(x => x.Assembly).ToArray();
             var buildSettings = UnityTestAssemblyBuilder.GetNUnitTestBuilderSettings(m_TestPlatform);
-            var testAssemblyBuilder = m_testAssemblyBuilderFactory(testJobData.executionSettings.orderedTestNames);
+            var testAssemblyBuilder = m_testAssemblyBuilderFactory(testJobData.executionSettings.orderedTestNames, testJobData.executionSettings.randomOrderSeed);
             var enumerator = testAssemblyBuilder.BuildAsync(assemblies, Enumerable.Repeat(m_TestPlatform, assemblies.Length).ToArray(), buildSettings);
             while (enumerator.MoveNext())
             {
@@ -53,7 +53,7 @@ namespace UnityEditor.TestTools.TestRunner.TestRun.Tasks
             {
                 throw new Exception("Test list not retrieved.");
             }
-            
+
             testJobData.testTree = testList;
             m_CallbacksDelegator.TestTreeRebuild(testList);
         }