+++ UnityEditor.TestRunner/TestRun/Tasks/LegacyPlayerRunTask.cs
@@ -11,7 +11,7 @@ namespace UnityEditor.TestTools.TestRunner.TestRun.Tasks
         public override IEnumerator Execute(TestJobData testJobData)
         {
             var executionSettings = testJobData.executionSettings;
-            var settings = PlaymodeTestsControllerSettings.CreateRunnerSettings(executionSettings.filters.Select(filter => filter.ToRuntimeTestRunnerFilter(executionSettings.runSynchronously)).ToArray(), testJobData.executionSettings.orderedTestNames, testJobData.executionSettings.featureFlags, executionSettings.retryCount, executionSettings.repeatCount, IsAutomated());
+            var settings = PlaymodeTestsControllerSettings.CreateRunnerSettings(executionSettings.filters.Select(filter => filter.ToRuntimeTestRunnerFilter(executionSettings.runSynchronously)).ToArray(), testJobData.executionSettings.orderedTestNames, testJobData.executionSettings.randomOrderSeed, testJobData.executionSettings.featureFlags, executionSettings.retryCount, executionSettings.repeatCount, IsAutomated());
             var launcher = new PlayerLauncher(settings, executionSettings.targetPlatform, executionSettings.overloadTestRunSettings, executionSettings.playerHeartbeatTimeout, executionSettings.playerSavePath);
             launcher.Run();