+++ UnityEngine.TestRunner/NUnitExtensions/UnityTestAssemblyBuilder.cs
@@ -14,10 +14,11 @@ namespace UnityEngine.TestTools.NUnitExtensions
     {
         private readonly string m_ProductName;
         private readonly ITestSuiteModifier[] m_TestSuiteModifiers;
-        public UnityTestAssemblyBuilder(string[] orderedTestNames)
+
+        public UnityTestAssemblyBuilder(string[] orderedTestNames, int randomSeed)
         {
-            m_TestSuiteModifiers = orderedTestNames != null && orderedTestNames.Length > 0
-                ? new ITestSuiteModifier[] {new OrderedTestSuiteModifier(orderedTestNames)}
+            m_TestSuiteModifiers = (orderedTestNames != null && orderedTestNames.Length > 0) || randomSeed != 0
+                ? new ITestSuiteModifier[] {new OrderedTestSuiteModifier(orderedTestNames, randomSeed)}
                 : new ITestSuiteModifier[0];
             m_ProductName = Application.productName;
         }
@@ -40,7 +41,7 @@ namespace UnityEngine.TestTools.NUnitExtensions
             {
                 var assembly = assemblies[index];
                 var platform = testPlatforms[index];
-                
+
                 var assemblySuite = Build(assembly, options) as TestSuite;
                 if (assemblySuite != null && assemblySuite.HasChildren)
                 {
@@ -50,7 +51,7 @@ namespace UnityEngine.TestTools.NUnitExtensions
 
                 yield return null;
             }
-            
+
             suite.ParseForNameDuplicates();
             suite.Properties.Set("platform", testPlatforms.MergeFlags());